Split util file into two for better treeshaking
