+++ src/components/pages/auth/registration.tsx
@@ -8,11 +8,8 @@ import { useRouter } from 'next/router'
 import nProgress from 'nprogress'
 import { useEffect, useState } from 'react'
 
-import {
-  verificationUrl,
-  VALIDATION_ERROR_TYPE,
-  changeButtonTypeOfSSOProvider,
-} from './utils'
+import { changeButtonTypeOfSSOProvider } from './ory-transforms'
+import { verificationUrl, VALIDATION_ERROR_TYPE } from './utils'
 import { kratos } from '@/auth/kratos'
 import { useCheckInstance } from '@/auth/use-check-instance'
 import { Flow, FlowType, handleFlowError } from '@/components/auth/flow'
